Remove Firebase secrets and add example configs
