+++ .gitignore
@@ -51,3 +51,6 @@ lib/firebase_options.dart
 
 # Private dependencies
 pubspec.lock
+# Firebase configuration files
+android/app/google-services.json
+ios/Runner/GoogleService-Info.plist
