+++ ios/Runner/GoogleService-Info.example.plist
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+    <key>CLIENT_ID</key>
+    <string>YOUR_CLIENT_ID</string>
+    <key>REVERSED_CLIENT_ID</key>
+    <string>YOUR_REVERSED_CLIENT_ID</string>
+    <key>API_KEY</key>
+    <string>YOUR_IOS_API_KEY</string>
+    <key>GCM_SENDER_ID</key>
+    <string>YOUR_SENDER_ID</string>
+    <key>PLIST_VERSION</key>
+    <string>1</string>
+    <key>BUNDLE_ID</key>
+    <string>your.bundle.id</string>
+    <key>PROJECT_ID</key>
+    <string>YOUR_PROJECT_ID</string>
+    <key>STORAGE_BUCKET</key>
+    <string>YOUR_BUCKET</string>
+</dict>
+</plist>
