--- ios/Runner/GoogleService-Info.plist
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CLIENT_ID</key>
-	<string>774086218053-jknu4giv8bvnuomngf8eifr516met4tr.apps.googleusercontent.com</string>
-	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.774086218053-jknu4giv8bvnuomngf8eifr516met4tr</string>
-	<key>API_KEY</key>
-	<string>AIzaSyCuiCbP94ir_yAV8BWiupKg0hEiLwi8CaU</string>
-	<key>GCM_SENDER_ID</key>
-	<string>774086218053</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.lween.app.lween</string>
-	<key>PROJECT_ID</key>
-	<string>lween-6e0ee</string>
-	<key>STORAGE_BUCKET</key>
-	<string>lween-6e0ee.appspot.com</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:774086218053:ios:8a66ba91e3d919b511ee2c</string>
-</dict>
-</plist>
